Improved email validation
