+++ packages/shared/src/lib/emails.shared.test.ts
@@ -1,0 +1,25 @@
+import {isValidEmail} from '@shared/lib/emails.shared';
+
+describe('isValidEmail', () => {
+  test('should return true for valid email addresses', () => {
+    expect(isValidEmail('test@example.com')).toBe(true);
+    expect(isValidEmail('user.name@domain.co.uk')).toBe(true);
+    expect(isValidEmail('user+tag@example.com')).toBe(true);
+  });
+
+  test('should return false for invalid email addresses', () => {
+    expect(isValidEmail('not-an-email')).toBe(false);
+    expect(isValidEmail('missing@domain')).toBe(false);
+    expect(isValidEmail('@nodomain.com')).toBe(false);
+    expect(isValidEmail('no@domain.')).toBe(false);
+    expect(isValidEmail('')).toBe(false);
+  });
+
+  test('should return false for non-string values', () => {
+    expect(isValidEmail(null)).toBe(false);
+    expect(isValidEmail(undefined)).toBe(false);
+    expect(isValidEmail(123)).toBe(false);
+    expect(isValidEmail({})).toBe(false);
+    expect(isValidEmail([])).toBe(false);
+  });
+});
