+++ packages/shared/src/lib/emails.shared.ts
@@ -1,0 +1,12 @@
+import type {EmailAddress} from '@shared/types/emails.types';
+
+// Source of regex: https://emailregex.com/index.html.
+const EMAIL_REGEX =
+  /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+
+/**
+ * Checks if a value is a valid `EmailAddress`.
+ */
+export function isValidEmail(maybeEmail: unknown): maybeEmail is EmailAddress {
+  return typeof maybeEmail === 'string' && EMAIL_REGEX.test(maybeEmail);
+}
