+++ packages/shared/src/lib/utils.shared.test.ts
@@ -7,10 +7,8 @@ import {
   filterNull,
   filterUndefined,
   formatWithCommas,
-  isDate,
   isInteger,
   isPositiveInteger,
-  isValidEmail,
   isValidPort,
   makeUuid,
   noopTrue,
@@ -74,30 +72,6 @@ describe('filterUndefined', () => {
 
   test('should handle empty array', () => {
     expect(filterUndefined([])).toEqual([]);
-  });
-});
-
-describe('isValidEmail', () => {
-  test('should return true for valid email addresses', () => {
-    expect(isValidEmail('test@example.com')).toBe(true);
-    expect(isValidEmail('user.name@domain.co.uk')).toBe(true);
-    expect(isValidEmail('user+tag@example.com')).toBe(true);
-  });
-
-  test('should return false for invalid email addresses', () => {
-    expect(isValidEmail('not-an-email')).toBe(false);
-    expect(isValidEmail('missing@domain')).toBe(false);
-    expect(isValidEmail('@nodomain.com')).toBe(false);
-    expect(isValidEmail('no@domain.')).toBe(false);
-    expect(isValidEmail('')).toBe(false);
-  });
-
-  test('should return false for non-string values', () => {
-    expect(isValidEmail(null)).toBe(false);
-    expect(isValidEmail(undefined)).toBe(false);
-    expect(isValidEmail(123)).toBe(false);
-    expect(isValidEmail({})).toBe(false);
-    expect(isValidEmail([])).toBe(false);
   });
 });
 
@@ -182,31 +156,6 @@ describe('partition', () => {
     const [numbers, strings] = partition(mixed, (item) => typeof item === 'number');
     expect(numbers).toEqual([1, 2, 3]);
     expect(strings).toEqual(['a', 'b']);
-  });
-});
-
-describe('isDate', () => {
-  test('should return true for Date objects', () => {
-    expect(isDate(new Date())).toBe(true);
-    expect(isDate(new Date('2023-01-01'))).toBe(true);
-  });
-
-  test('should return false for non-Date objects', () => {
-    expect(isDate(null)).toBe(false);
-    expect(isDate(undefined)).toBe(false);
-    expect(isDate('2023-01-01')).toBe(false);
-    expect(isDate(123)).toBe(false);
-    expect(isDate({})).toBe(false);
-    expect(isDate([])).toBe(false);
-    expect(isDate(true)).toBe(false);
-  });
-
-  test('should return false for Date-like objects that are not actual Date instances', () => {
-    const dateLike = {
-      getTime: () => 1672531200000,
-      toISOString: () => '2023-01-01T00:00:00.000Z',
-    };
-    expect(isDate(dateLike)).toBe(false);
   });
 });
